Atari 2600 Ms Pacman OpenCV Method
This method is used for detecting gold points for the agent in Atar Ms. Pacman gaming environment. This method was implemented by applying OpenCV APIs and it was composed in Python. By utilizing this method, the agent can efficiently detect its surrounding gold points and therefore, improves the training efficiency of the emotion-based DQN model by 35%.